+++ lib/vips/compass_direction.rb
@@ -2,16 +2,16 @@ module Vips
 
     # A direction on a compass used for placing images. See {Image#gravity}.
     #
-    # * `"centre"`
-    # * `"north"`
-    # * `"east"`
-    # * `"south"`
-    # * `"west"`
-    # * `"north-east"`
-    # * `"south-east"`
-    # * `"south-west"`
-    # * `"north-west"`
+    # * `:centre`
+    # * `:north`
+    # * `:east`
+    # * `:south`
+    # * `:west`
+    # * `:north-east`
+    # * `"south-east`
+    # * `:south-west`
+    # * `:north-west`
 
-    class CompassDirection < String
+    class CompassDirection < Symbol
     end
 end