+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * add support for uint64 parameters
 * add `draw_point` convenience method
 * add docs for `CompassDirection` [janko-m]
+* add `MAX_COORD` constant
 
 ## Version 2.0.9 (2017-12-21)